refactor: update identifier retrieval to use coreauth package
Refactored the identifier retrieval logic in context.go and context_test.go to utilize the coreauth package instead of the extensions/auth package. This change simplifies the code by removing the need for an authenticator parameter in the identifierFromKey function, enhancing clarity and maintainability. Updated related tests to reflect this change and ensure consistent behavior in node identification.
